feat(import): update validation and pagination logic in ImportFailureController and TransactionController
- Adjusted validation rules for transaction creation to allow any string for the 'type' field and made 'balance_after_transaction' required.
- Enhanced pagination logic in ImportFailureController to dynamically set the number of records displayed based on filters.
- Removed deprecated transaction creation logic from the frontend, streamlining the review process in the ImportFailures component.
- Improved the ReviewInterface to include a new transaction creation handler, enhancing user feedback during the import failure review.
